requirements.txt
python-docx==0.8.11
requests==2.31.0
pypinyin==0.49.0 